moved pages folder outside components folder to src and fixed paths for components/pages
