+++ client/src/pages/Home.js
@@ -1,6 +1,6 @@
 import React, { Component} from "react";
-import "../../App.css";
-import Profile from "../Profile";
+import "../App.css";
+import Profile from "../components/Profile";
 
 
 class Home extends Component {
