+++ config/config.json
@@ -1,7 +1,7 @@
 {
   "development": {
     "username": "root",
-    "password": "password",
+    "password": "",
     "database": "employees_db",
     "host": "localhost",
     "dialect": "mysql"
